bitweaver perf test

+++ src/scan/bitpack/BitWeaverHScanner128.cpp
@@ -8,7 +8,7 @@
 #define WORD_IN_SIMD 2
 
 
-__m128i make128(int num, int entrySize) {
+__m128i make128(uint64_t num, int entrySize) {
     int entryInWord = 64 / (entrySize + 1);
     uint64_t word = 0;
     for (int i = 0; i < entryInWord; i++) {
