+++ src/scan/bitpack/BitWeaverHScanner256.cpp
@@ -9,7 +9,7 @@
 #define WORD_IN_SIMD 4
 
 
-__m256i make256(int num, int entrySize) {
+__m256i make256(uint64_t num, int entrySize) {
     int entryInWord = 64 / (entrySize + 1);
     uint64_t word = 0;
     for (int i = 0; i < entryInWord; i++) {
